requirements.txt
discord.py==1.7.3
pymongo==3.12.0
python-dotenv~=0.19.0
requests~=2.26.0
